api/requirements.txt
fastapi
uvicorn
pandas
sentence-transformers
pyvespa
